+++ src/main/java/com/emre/vetproject/business/abstracts/ICustomerService.java
@@ -1,6 +1,5 @@
 package com.emre.vetproject.business.abstracts;
 
-import com.emre.vetproject.dto.response.customer.CustomerResponse;
 import com.emre.vetproject.model.Animal;
 import com.emre.vetproject.model.Customer;
 import org.springframework.data.domain.Page;
@@ -9,13 +8,17 @@ import java.util.List;
 
 public interface ICustomerService {
     Customer save(Customer customer);
-    Customer get(int id);
+
+    Customer get(long id);
+
     Page<Customer> cursor(int page, int pageSize);
 
     Customer update(Customer customer);
 
-    boolean delete(int id);
-    List<Animal> getAllAnimalsByCustomer(int customerId);
+    boolean delete(long id);
+
+    List<Animal> getAllAnimalsByCustomer(long customerId);
+
     List<Customer> findCustomersByName(String name);
 
 }
