+++ src/main/java/com/emre/vetproject/business/abstracts/IVaccineService.java
@@ -3,7 +3,6 @@ package com.emre.vetproject.business.abstracts;
 import com.emre.vetproject.dto.request.vaccine.VaccineDateSearchRequest;
 import com.emre.vetproject.dto.request.vaccine.VaccineSaveRequest;
 import com.emre.vetproject.model.Animal;
-import com.emre.vetproject.model.Customer;
 import com.emre.vetproject.model.Vaccine;
 import org.springframework.data.domain.Page;
 
@@ -12,17 +11,20 @@ import java.util.Set;
 
 public interface IVaccineService {
     Vaccine save(Vaccine vaccine);
+
     void checkVaccine(Animal animal, VaccineSaveRequest vaccineSaveRequest);
-    Vaccine get(int id);
-    Page<Vaccine> cursor(Animal animal,int page, int pageSize);
+
+    Vaccine get(long id);
+
+    Page<Vaccine> cursor(Animal animal, int page, int pageSize);
 
     Set<Animal> searchByDateRange(VaccineDateSearchRequest vaccineDateSearchRequest);
 
     Vaccine update(Vaccine vaccine);
 
-    boolean delete(int id);
+    boolean delete(long id);
 
-    List<Animal> getAllAnimalsByVaccine(int vaccineId);
+    List<Animal> getAllAnimalsByVaccine(long vaccineId);
 
 
 }
