+++ src/main/java/com/emre/vetproject/dao/ICustomerRepo.java
@@ -1,7 +1,5 @@
 package com.emre.vetproject.dao;
 
-import com.emre.vetproject.dto.response.customer.CustomerResponse;
-import com.emre.vetproject.model.Animal;
 import com.emre.vetproject.model.Customer;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
@@ -9,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ICustomerRepo extends JpaRepository<Customer, Integer> {
+public interface ICustomerRepo extends JpaRepository<Customer, Long> {
     List<Customer> findCustomersByName(String name);
 }
