+++ src/main/java/com/emre/vetproject/dto/request/animal/AnimalSaveRequest.java
@@ -1,7 +1,6 @@
 package com.emre.vetproject.dto.request.animal;
 
 import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Positive;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -12,7 +11,7 @@ import java.time.LocalDate;
 @AllArgsConstructor
 @NoArgsConstructor
 public class AnimalSaveRequest {
-    private int id;
+    private Long id;
     @NotNull(message = "isim boş bırakılamaz!")
     private String name;
     @NotNull(message = "tür boş bırakılamaz!")
