+++ src/main/java/com/emre/vetproject/dto/response/vaccine/VaccineResponse.java
@@ -1,19 +1,16 @@
 package com.emre.vetproject.dto.response.vaccine;
 
-import com.emre.vetproject.model.Animal;
-import jakarta.persistence.OneToMany;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import java.time.LocalDate;
-import java.util.List;
 
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class VaccineResponse {
-    private int id;
+    private Long id;
     private String name;
     private String code;
     private LocalDate protectionStartDate;
